+++ src/Process/BootstrapProcess.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\ParallelProcess\Process;
+
+class BootstrapProcess extends Process implements BootstrapProcessInterface
+{
+}